0) Может читать/писать(скоро) данные внутреней памяти Панели Оператора. Для этого ini уставок панели зашивается в ресурсы и теперь к своим параметрам обращаться "SLF/FLASH/Tag/" 1) config.json добавил внутренние параметры дисплея 2) умеет 0x16 Mask Write Register команды Modbus для записи битовых и байтовых данных 3) Добавлена странца настроек панели PagePanelSettings 4) Добавлена страница калибровок PageCalibration 5) Страница PageParameterListEdit для редактирования списка TPrmList

+++ App/src/MCU/Helpers/crc16.c
@@ -71,7 +71,7 @@ const u8 CRC16Lo[] = {
 0x82, 0x42, 0x43, 0x83, 0x41, 0x81, 0x80, 0x40
 } ;
 
-u16 crc16(u8 *puchMsg, u32 DataLen)
+u16 crc16(u8 *puchMsg, u16 DataLen)
 {
   u8 CRCHi ; /* high byte of CRC initialized */
   u8 CRCLo ; /* low byte of CRC initialized */
